requirements.txt
pandas
scikit-learn
numpy
matplotlib
seaborn
mysql-connector-python
pymysql
python-dotenv